requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pydantic==2.8.2
pyyaml==6.0.2
tldextract==5.1.2
playwright==1.46.0
tenacity==8.5.0
